+++ components/providers/AuthFlagProvider.jsx
@@ -1,14 +1,21 @@
-import { createContext } from "react";
+import { createContext, useEffect, useState } from "react";
+import { getAuth, onAuthStateChanged } from 'firebase/auth';
 
 export const AuthFlagContext = createContext({});
 
-export const AuthFlagProvider = prps => {
-    const { children } = prps;
+export const AuthFlagProvider = ({ children }) => {
+    const [isAuth, setIsAuth] = useState(false);
 
-    const sampleObj = { sampleValue: 'テスト' };
+    useEffect(() => {
+        const auth = getAuth();
+        const unsubscribe = onAuthStateChanged(auth, user => {
+            setIsAuth(!!user);
+        });
+        return () => unsubscribe();
+    }, []);
 
     return (
-        <AuthFlagContext.Provider value={sampleObj}>
+        <AuthFlagContext.Provider value={{ isAuth, setIsAuth }}>
             {children}
         </AuthFlagContext.Provider>
     );
